Increase all timeouts for cinder (haproxy+httpd+RPC)
This is useful for example when cinder is used as a glance backend
and the creation of the image takes quite some time, as the may hit
resource limits (it may be overloaded or simply small for testing purposes).

Also, remove the previous haproxy override (which seems to work,
but apiTimeout applies to more settings).

Glance got a similar timeout increase in a previous change.